requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
pytube==15.0.0
pillow==10.0.1